refactor: rename into yakusha
