+++ README.md
@@ -1,12 +1,12 @@
-# Kumo JSON
+# Yakusha
 
-[![latest version](https://img.shields.io/github/v/release/ichiro-its/kumo_json)](https://github.com/ichiro-its/kumo_json/releases/)
-[![commits since latest version](https://img.shields.io/github/commits-since/ichiro-its/kumo_json/latest)](https://github.com/ichiro-its/kumo_json/commits/master)
-[![milestone](https://img.shields.io/github/milestones/progress/ichiro-its/kumo_json/1?label=milestone)](https://github.com/ichiro-its/kumo_json/milestone/1)
-[![license](https://img.shields.io/github/license/ichiro-its/kumo_json)](./LICENSE)
-[![test status](https://img.shields.io/github/workflow/status/ichiro-its/kumo_json/Build%20and%20Test?label=test)](https://github.com/ichiro-its/kumo_json/actions)
+[![latest version](https://img.shields.io/github/v/release/ichiro-its/yakusha)](https://github.com/ichiro-its/yakusha/releases/)
+[![commits since latest version](https://img.shields.io/github/commits-since/ichiro-its/yakusha/latest)](https://github.com/ichiro-its/yakusha/commits/master)
+[![milestone](https://img.shields.io/github/milestones/progress/ichiro-its/yakusha/1?label=milestone)](https://github.com/ichiro-its/yakusha/milestone/1)
+[![license](https://img.shields.io/github/license/ichiro-its/yakusha)](./LICENSE)
+[![test status](https://img.shields.io/github/workflow/status/ichiro-its/yakusha/Build%20and%20Test?label=test)](https://github.com/ichiro-its/yakusha/actions)
 
-This package provides a [ROS 2 interface](https://docs.ros.org/en/foxy/Concepts/About-ROS-Interfaces.html) to [JSON](https://www.json.org/) serialization library for [Kumo](https://github.com/ichiro-its/kumo) WebSocket to ROS 2 bridge.
+This package provides a [ROS 2 interface](https://docs.ros.org/en/foxy/Concepts/About-ROS-Interfaces.html) to [JSON](https://www.json.org/) serialization library.
 
 ## License
