--- kumo_json/setup.cfg
@@ -1,4 +1,0 @@
-[develop]
-script-dir=$base/lib/kumo_json
-[install]
-install-scripts=$base/lib/kumo_json
