+++ yakusha/package.xml
@@ -1,18 +1,18 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>kumo_json</name>
+  <name>yakusha</name>
   <version>0.1.0</version>
-  <description>ROS 2 interface to JSON serialization for Kumo package</description>
+  <description>ROS 2 interface for JSON serialization project</description>
   <maintainer email="alfi.maulana.f@gmail.com">Alfi Maulana</maintainer>
   <license>MIT License</license>
-  <url>https://github.com/ichiro-its/kumo_json</url>
+  <url>https://github.com/ichiro-its/yakusha</url>
   <depend>rclpy</depend>
   <test_depend>ament_copyright</test_depend>
   <test_depend>ament_flake8</test_depend>
   <test_depend>ament_pep257</test_depend>
   <test_depend>std_msgs</test_depend>
-  <test_depend>kumo_json_interfaces</test_depend>
+  <test_depend>yakusha_interfaces</test_depend>
   <test_depend>python3-pytest</test_depend>
   <export>
     <build_type>ament_python</build_type>
