+++ yakusha/setup.cfg
@@ -1,0 +1,4 @@
+[develop]
+script-dir=$base/lib/yakusha
+[install]
+install-scripts=$base/lib/yakusha
