+++ yakusha/test/test_copyright.py
