+++ yakusha/test/test_serialization/test_complex/test_non_floats_msg.py
@@ -18,9 +18,9 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 # THE SOFTWARE.
 
-from kumo_json_interfaces.msg import NonFloats
+from yakusha_interfaces.msg import NonFloats
 
-from kumo_json import msg_to_json, json_to_msg
+from yakusha import msg_to_json, json_to_msg
 
 
 def test_non_floats_msg():
