+++ yakusha_interfaces/msg/arrays/Arrays.msg
