+++ yakusha_interfaces/msg/arrays/BoundedDynamicArrays.msg
