+++ yakusha_interfaces/msg/primitives/NonNumbers.msg
