+++ yakusha_interfaces/package.xml
@@ -1,12 +1,12 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>kumo_json_interfaces</name>
+  <name>yakusha_interfaces</name>
   <version>0.1.0</version>
-  <description>ROS 2 interfaces for Kumo JSON package</description>
+  <description>ROS 2 interface serialization project</description>
   <maintainer email="alfi.maulana.f@gmail.com">Alfi Maulana</maintainer>
   <license>MIT License</license>
-  <url>https://github.com/ichiro-its/kumo_json</url>
+  <url>https://github.com/ichiro-its/yakusha</url>
   <buildtool_depend>ament_cmake</buildtool_depend>
   <buildtool_depend>rosidl_default_generators</buildtool_depend>
   <test_depend>ament_lint_auto</test_depend>
